requirements.txt
Django==2.2.5
psycopg2-binary==2.8.3
requests==2.22.0
djangorestframework==3.10.2
freezegun==0.3.12
django-crispy-forms==1.7.1
django-allauth==0.38.0
django_cron
